requirements.txt
# Azure Compliance Collector Dependencies
#
# We use direct REST API calls (via requests) for most collectors rather than
# the azure-mgmt-* SDK clients. This avoids breaking changes between SDK major
# versions which have caused repeated import errors.
#
# The SDK clients we still use:
#   azure-identity        — DefaultAzureCredential (handles OIDC, CLI, managed identity)
#   azure-mgmt-authorization — RBAC role assignments (stable, used in collect_entra_id.py)
#
# Everything else goes through requests + bearer token directly against the
# Azure REST API and Microsoft Graph REST API.

azure-identity>=1.15.0
azure-mgmt-authorization>=4.0.0
requests>=2.31.0

# Report generation
pyyaml>=6.0
